+++ cqrs-core/CHANGELOG.md
@@ -1,6 +1,7 @@
 # master
 
-* No changes yet
+* Breaking change to `EventSource::Events` so that the iterated items are no
+  longer results.
 
 # [[0.1.0] 2019-02-01](https://github.com/cq-rs/cqrs/releases/tag/cqrs-core-0.1.0)